+++ .storybook/config.js
@@ -1,0 +1,9 @@
+import { configure } from '@storybook/react';
+
+import '../src/style.css';
+
+function loadStories() {
+  require('../stories');
+}
+
+configure(loadStories, module);